Returnerer expired cache item fra response cache dersom ny fetch feiler
